Update Employee.java
Actualziación de entity employee

+++ src/main/java/com/example/tarea1_20220270_20210845/entity/Employee.java
@@ -28,7 +28,7 @@ public class Employee {
     @JoinColumn(name = "job_id")
     private Job job;
 
-    private BigDecimal salary;
+    private Double salary;
 
     @ManyToOne
     @JoinColumn(name = "manager_id")
